Rename replaceLegacyCovidExplorer → replaceExplorerRedirects
