+++ baker/formatWordpressPost.tsx
@@ -36,7 +36,7 @@ import { SVG } from "mathjax-full/js/output/svg"
 import { liteAdaptor } from "mathjax-full/js/adaptors/liteAdaptor"
 import { RegisterHTMLHandler } from "mathjax-full/js/handlers/html"
 import { AllPackages } from "mathjax-full/js/input/tex/AllPackages"
-import { replaceIframesWithExplorerRedirectsInWordPressPost } from "./replaceLegacyCovidExplorer"
+import { replaceIframesWithExplorerRedirectsInWordPressPost } from "./replaceExplorerRedirects"
 import { EXPLORERS_ROUTE_FOLDER } from "../explorer/ExplorerConstants"
 
 const initMathJax = () => {
